Refactor payload fields in ConnectionOpenConfirm
